Stabilize tests, Separate classes, add comments, clean up code
