+++ V2/kafka/src/main/java/com/neverwinterdp/kafka/tool/KafkaTool.java
@@ -12,7 +12,6 @@ import java.util.HashMap;
 import java.util.Iterator;
 import java.util.List;
 import java.util.Map;
-import java.util.Properties;
 import java.util.Random;
 
 import kafka.admin.AdminUtils;
@@ -20,10 +19,7 @@ import kafka.admin.PreferredReplicaLeaderElectionCommand;
 import kafka.admin.ReassignPartitionsCommand;
 import kafka.admin.TopicCommand;
 import kafka.admin.TopicCommand.TopicCommandOptions;
-import kafka.api.PartitionOffsetRequestInfo;
 import kafka.common.TopicAndPartition;
-import kafka.javaapi.OffsetRequest;
-import kafka.javaapi.OffsetResponse;
 import kafka.javaapi.PartitionMetadata;
 import kafka.javaapi.TopicMetadata;
 import kafka.javaapi.TopicMetadataRequest;
@@ -46,16 +42,16 @@ import com.neverwinterdp.kafka.BrokerRegistration;
 public class KafkaTool implements Closeable {
   private String name;
   private String zkConnects;
-  private ZooKeeper zkClient ;
-  private SimpleConsumer consumer ;
+  private ZooKeeper zkClient;
+  private SimpleConsumer consumer;
 
   public KafkaTool(String name, String zkConnects) {
     this.name = name;
     this.zkConnects = zkConnects;
   }
-  
+
   public void connect() throws Exception {
-    Watcher watcher = new  Watcher() {
+    Watcher watcher = new Watcher() {
       @Override
       public void process(WatchedEvent event) {
       }
@@ -63,51 +59,76 @@ public class KafkaTool implements Closeable {
     zkClient = new ZooKeeper(zkConnects, 15000, watcher);
     nextLeader();
   }
-  
+
   public void connect(String zkConnect) throws Exception {
     this.zkConnects = zkConnect;
     connect();
   }
-  
-  public ZooKeeper getZookeeper() { return this.zkClient ;}
-  
+
+  public ZooKeeper getZookeeper() {
+    return this.zkClient;
+  }
+
   @Override
   public void close() throws IOException {
-    if(consumer != null) {
+    if (consumer != null) {
       consumer.close();
-      consumer = null ;
+      consumer = null;
     }
-    if(zkClient != null) {
+    if (zkClient != null) {
       try {
         zkClient.close();
       } catch (InterruptedException e) {
         throw new IOException(e);
       }
-      zkClient = null ;
+      zkClient = null;
     }
   }
-  
-  public void createTopic(String topicName, int numOfReplication, int numPartitions)  {
-    // Create a ZooKeeper client
+
+  public void createTopic(String topicName, int numOfReplication, int numPartitions) throws Exception {
+    String[] args = { "--create",
+        "--partition", String.valueOf(numPartitions),
+        "--replication-factor", String.valueOf(numOfReplication),
+        "--topic", topicName,
+        "--zookeeper", zkConnects
+    };
+
+    createTopic(args);
+  }
+/**
+ * Create a topic. 
+ * For valid configs see https://cwiki.apache.org/confluence/display/KAFKA/Replication+tools#Replicationtools-Howtousethetool?.3
+ *
+ * @See https://kafka.apache.org/documentation.html#topic-config 
+ * for more valid configs
+ * */
+  public void createTopic(String[] args) throws Exception {
     int sessionTimeoutMs = 10000;
     int connectionTimeoutMs = 10000;
-    ZkClient zkClient = new ZkClient(zkConnects, sessionTimeoutMs, connectionTimeoutMs, ZKStringSerializer$.MODULE$);
-    // Create a topic named "myTopic" with 8 partitions and a replication factor of 3
-    Properties topicConfig = new Properties();
-    AdminUtils.createTopic(zkClient, topicName, numPartitions, numOfReplication, topicConfig);
+
+    TopicCommandOptions options = new TopicCommandOptions(args);
+    ZkClient client = new ZkClient(zkConnects, sessionTimeoutMs, connectionTimeoutMs, ZKStringSerializer$.MODULE$);
+    if (topicExits(name)) {
+      TopicCommand.deleteTopic(client, options);
+    }
+
+    TopicCommand.createTopic(client, options);
     try {
       Thread.sleep(3000);
     } catch (InterruptedException e) {
       e.printStackTrace();
     }
-    zkClient.close();
+    client.close();
   }
-  
+
   /**
    * This delete method doesn't work
+   *
    * @param topicName
    * @throws Exception
    */
+  //TODO un-deprecate by using TopicCommand.deleteTopic()
+  //Also ensure cluster has "delete.topic.enable" = "true"
   @Deprecated
   public void deleteTopic(String topicName) throws Exception {
     int sessionTimeoutMs = 1000;
@@ -116,40 +137,41 @@ public class KafkaTool implements Closeable {
     AdminUtils.deleteTopic(zkClient, topicName);
     zkClient.close();
   }
-  
+
   public boolean topicExits(String topicName) throws Exception {
     int sessionTimeoutMs = 1000;
     int connectionTimeoutMs = 1000;
     ZkClient zkClient = new ZkClient(zkConnects, sessionTimeoutMs, connectionTimeoutMs, ZKStringSerializer$.MODULE$);
-    boolean exists = AdminUtils.topicExists(zkClient, topicName) ;
+    boolean exists = AdminUtils.topicExists(zkClient, topicName);
     zkClient.close();
     return exists;
   }
-  
-  public String getKafkaBrokerList() throws KeeperException, InterruptedException  {
+
+  public String getKafkaBrokerList() throws KeeperException, InterruptedException {
     StringBuilder b = new StringBuilder();
     List<BrokerRegistration> registrations = getBrokerRegistration();
-    for(int i = 0; i < registrations.size(); i++) {
+    for (int i = 0; i < registrations.size(); i++) {
       BrokerRegistration registration = registrations.get(i);
-      if(i > 0) b.append(",");
+      if (i > 0)
+        b.append(",");
       b.append(registration.getHost()).append(":").append(registration.getPort());
     }
     return b.toString();
   }
-  
+
   public List<BrokerRegistration> getBrokerRegistration() throws KeeperException, InterruptedException {
-    List<String> ids = zkClient.getChildren("/brokers/ids", false) ;
+    List<String> ids = zkClient.getChildren("/brokers/ids", false);
     List<BrokerRegistration> holder = new ArrayList<BrokerRegistration>();
-    for(int i = 0; i < ids.size(); i++) {
+    for (int i = 0; i < ids.size(); i++) {
       String brokerId = ids.get(i);
-      BrokerRegistration registration = 
-        ZKTool.getDataAs(zkClient, "/brokers/ids/" + brokerId, BrokerRegistration.class);
+      BrokerRegistration registration =
+          ZKTool.getDataAs(zkClient, "/brokers/ids/" + brokerId, BrokerRegistration.class);
       registration.setBrokerId(brokerId);
       holder.add(registration);
     }
     return holder;
   }
-  
+
   public TopicMetadata findTopicMetadata(final String topic) throws Exception {
     Operation<TopicMetadata> findTopicOperation = new Operation<TopicMetadata>() {
       @Override
@@ -159,32 +181,33 @@ public class KafkaTool implements Closeable {
         TopicMetadataResponse resp = consumer.send(req);
 
         List<TopicMetadata> topicMetadatas = resp.topicsMetadata();
-        if(topicMetadatas.size() != 1) {
+        if (topicMetadatas.size() != 1) {
           throw new Exception("Expect to find 1 topic " + topic + ", but found " + topicMetadatas.size());
         }
-        
+
         return topicMetadatas.get(0);
       }
     };
     return findTopicOperation.execute();
   }
-  
+
   public PartitionMetadata findPartitionMetadata(String topic, int partition) throws Exception {
     TopicMetadata topicMetadata = findTopicMetadata(topic);
-    for(PartitionMetadata sel : topicMetadata.partitionsMetadata()) {
-      if(sel.partitionId() == partition) return sel;
+    for (PartitionMetadata sel : topicMetadata.partitionsMetadata()) {
+      if (sel.partitionId() == partition)
+        return sel;
     }
     return null;
   }
-  
+
   private void nextLeader() throws Exception {
     List<BrokerRegistration> registrations = getBrokerRegistration();
-    if(consumer != null) {
+    if (consumer != null) {
       //Remove the current select broker
       Iterator<BrokerRegistration> i = registrations.iterator();
-      while(i.hasNext()) {
+      while (i.hasNext()) {
         BrokerRegistration sel = i.next();
-        if(sel.getHost().equals(consumer.host()) && sel.getPort() == consumer.port()) {
+        if (sel.getHost().equals(consumer.host()) && sel.getPort() == consumer.port()) {
           i.remove();
           break;
         }
@@ -196,37 +219,14 @@ public class KafkaTool implements Closeable {
     BrokerRegistration registration = registrations.get(random.nextInt(registrations.size()));
     consumer = new SimpleConsumer(registration.getHost(), registration.getPort(), 100000, 64 * 1024, name /*clientId*/);
   }
-  
-  static public long getLastOffset(SimpleConsumer consumer, String topic, int partition, long whichTime, String clientName) {
-    TopicAndPartition topicAndPartition = new TopicAndPartition(topic, partition);
-    Map<TopicAndPartition, PartitionOffsetRequestInfo> requestInfo = new HashMap<TopicAndPartition, PartitionOffsetRequestInfo>();
-    requestInfo.put(topicAndPartition, new PartitionOffsetRequestInfo(whichTime, 1));
-    OffsetRequest request = new OffsetRequest(requestInfo, kafka.api.OffsetRequest.CurrentVersion(), clientName);
-    OffsetResponse response = consumer.getOffsetsBefore(request);
 
-    if (response.hasError()) {
-      System.out.println("Error fetching data Offset Data the Broker. Reason: " + response.errorCode(topic, partition));
-      return 0;
-    }
-    long[] offsets = response.offsets(topic, partition);
-    return offsets[0];
-  }
-  
-  static interface Operation<T> {
-    public T execute() throws Exception;
-  }
-
-  public void createTopic(String[] args) {    
-    TopicCommandOptions options = new TopicCommandOptions(args);
-    ZkClient client = new ZkClient(zkConnects, 10000, 10000, ZKStringSerializer$.MODULE$);
-    TopicCommand.createTopic(client, options);
-
-    client.close();
-  }
+  // Move Leader to first broker in ISR
+  // https://kafka.apache.org/documentation.html#basic_ops_leader_balancing
+  //https://cwiki.apache.org/confluence/display/KAFKA/Replication+tools#Replicationtools-2.PreferredReplicaLeaderElectionTool
 
   public void moveLeaderToPreferredReplica(String topic, int partition) {
     ZkClient client = new ZkClient(zkConnects, 10000, 10000, ZKStringSerializer$.MODULE$);
-    
+
     TopicAndPartition topicAndPartition = new TopicAndPartition(topic, partition);
     //move leader to broker 1
     Set<TopicAndPartition> topicsAndPartitions = asScalaSet(Collections.singleton(topicAndPartition));
@@ -234,14 +234,21 @@ public class KafkaTool implements Closeable {
         topicsAndPartitions);
 
     commands.moveLeaderToPreferredReplica();
-    
     client.close();
-    
   }
 
+  /**
+   * Re-assign topic/partition to remainingBrokers
+   * Remaining brokers is a list of id's of the brokers where the topic/partition is to be moved to.
+   * 
+   *   Thus if remainingBrokers = [1,2] the topic will be moved to brokers 1 and 2 
+   *   
+   *   @see https://kafka.apache.org/documentation.html#basic_ops_cluster_expansion
+   *   @see https://cwiki.apache.org/confluence/display/KAFKA/Replication+tools#Replicationtools-6.ReassignPartitionsTool
+   * */
   public boolean reassignPartition(String topic, int partition, List<Object> remainingBrokers) {
     ZkClient client = new ZkClient(zkConnects, 10000, 10000, ZKStringSerializer$.MODULE$);
-    
+
     TopicAndPartition topicAndPartition = new TopicAndPartition(topic, partition);
 
     Buffer<Object> seqs = asScalaBuffer(remainingBrokers);
@@ -249,7 +256,11 @@ public class KafkaTool implements Closeable {
     map.put(topicAndPartition, seqs);
     scala.collection.mutable.Map<TopicAndPartition, Seq<Object>> x = asScalaMap(map);
     ReassignPartitionsCommand command = new ReassignPartitionsCommand(client, x);
-    
+
     return command.reassignPartitions();
-   }
+  }
+
+  static interface Operation<T> {
+    public T execute() throws Exception;
+  }
 }
